Fix doSave to return correct header on new record
Update PickerFieldEditHandler::doSave so that correct x-controllerurl headers are returned to front end so that Add New Record button in the lower right will work after creating a new record with the PickerField.  Without this fix, the URL is not updated to the edit URL of the new record so the Add new record does nothing since it thinks the page is already at the correct URL for creating a new record.